requirements.txt
# requirements.txt

absl-py==2.1.0
annotated-types==0.6.0
asgiref==3.8.1
astunparse==1.6.3
async==0.6.2
asyncio==3.4.3
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==42.0.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
dm-tree==0.1.8
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Uploads==0.2.1
flatbuffers==24.3.7
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.3.2
keras==3.1.1
langcodes==3.3.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.10
mysql-connector-python==8.3.0
namex==0.0.7
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pathlib_abc==0.1.1
pathy==0.11.0
preshed==3.0.9
protobuf==4.25.3
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.4.0
torch==2.2.2
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0